services/graphrag/pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "Graph retrieval and RAG query service"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    # Graph & databases
    "falkordb>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",
    # LangChain / LlamaIndex
    "langchain-core>=0.2.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "llama-index-llms-groq>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    # Observability
    "langfuse>=2.45.0",
    "openinference-instrumentation-llama-index>=2.0.0",
    # MCP
    "mcp>=1.0.0",
    # ML / NLP
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "tiktoken>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
